[Content] Rename storage -> traits to resolve module self-reference
