+++ monas-content/src/infrastructure/storage/traits.rs
